refactor: extract the tool-step display vocabulary into a leaf module
Move the tool icon map and the tool-name-to-prose tables out of StreamEntry into a dependency-free toolDisplay module. The wording is about to gain a second consumer in the chat composer's loading indicator, and importing it from a component that pulls in the markdown renderer and the whole card set is both an awkward dependency and untestable.

Behaviour-preserving: the functions move verbatim, and StreamEntry imports what it already used. Add the unit tests the move makes possible, covering the MCP namespace parsing, the branches whose wording changes with the payload, and the unknown-tool fallbacks. They run under the node environment, since the package's jsdom setup currently fails to load.
